feature: moved 'not_processed_burn_tickets' endpoint to sharders
