+++ code/go/0chain.net/chaincore/chain/state_handler_test.go
@@ -183,169 +183,6 @@ func TestChain_GetBalanceHandler(t *testing.T) {
 	}
 }
 
-func TestChain_GetNotProcessedBurnTicketsHandler(t *testing.T) {
-	lfb := block.NewBlock("", 1)
-	lfb.ClientState = util.NewMerklePatriciaTrie(util.NewMemoryNodeDB(), 1, nil)
-	serverChain := chain.NewChainFromConfig()
-	serverChain.LatestFinalizedBlock = lfb
-
-	eventDb, err := event.NewInMemoryEventDb(config.DbAccess{}, config.DbSettings{
-		Debug:                 true,
-		PartitionChangePeriod: 1,
-	})
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		err = eventDb.Drop()
-		require.NoError(t, err)
-
-		eventDb.Close()
-	})
-
-	serverChain.EventDb = eventDb
-
-	sctx := serverChain.GetStateContextI()
-	lfb.ClientState = sctx.GetState()
-
-	tests := []struct {
-		name string
-		body func(t *testing.T)
-	}{
-		{
-			name: "Get not processed burn tickets of the client, which hasn't performed any burn operations, should work",
-			body: func(t *testing.T) {
-				target := url.URL{Path: "/v1/client/get/not_processed_burn_tickets"}
-
-				query := target.Query()
-
-				query.Add("ethereum_address", ethereumAddress)
-				query.Add("client_id", clientID)
-				query.Add("nonce", "0")
-
-				target.RawQuery = query.Encode()
-
-				req := httptest.NewRequest(http.MethodGet, target.String(), nil)
-
-				respRaw, err := serverChain.GetNotProcessedBurnTicketsHandler(context.Background(), req)
-				require.NoError(t, err)
-
-				resp, ok := respRaw.([]*state.BurnTicket)
-				require.True(t, ok)
-				require.Len(t, resp, 0)
-			},
-		},
-		{
-			name: "Get not processed burn tickets of the client, which has performed burn operation, should work",
-			body: func(t *testing.T) {
-				err := eventDb.Get().Model(&event.BurnTicket{}).Create(&event.BurnTicket{
-					UserID:          clientID,
-					EthereumAddress: ethereumAddress,
-					Hash:            hash,
-					Nonce:           1,
-				}).Error
-				require.NoError(t, err)
-
-				target := url.URL{Path: "/v1/client/get/not_processed_burn_tickets"}
-
-				query := target.Query()
-
-				query.Add("ethereum_address", ethereumAddress)
-				query.Add("client_id", clientID)
-				query.Add("nonce", "0")
-
-				target.RawQuery = query.Encode()
-
-				req := httptest.NewRequest(http.MethodGet, target.String(), nil)
-
-				respRaw, err := serverChain.GetNotProcessedBurnTicketsHandler(context.Background(), req)
-				require.NoError(t, err)
-
-				resp, ok := respRaw.([]*state.BurnTicket)
-				require.True(t, ok)
-				require.Len(t, resp, 1)
-
-				err = eventDb.Get().Model(&event.BurnTicket{}).Where("user_id = ?", clientID).Delete(&event.BurnTicket{}).Error
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "Get not processed burn tickets not providing client id, should not work",
-			body: func(t *testing.T) {
-				target := url.URL{Path: "/v1/client/get/not_processed_burn_tickets"}
-
-				query := target.Query()
-
-				query.Add("ethereum_address", ethereumAddress)
-				query.Add("nonce", "0")
-
-				target.RawQuery = query.Encode()
-
-				req := httptest.NewRequest(http.MethodGet, target.String(), nil)
-
-				resp, err := serverChain.GetNotProcessedBurnTicketsHandler(context.Background(), req)
-				require.Error(t, err)
-				require.Nil(t, resp)
-			},
-		},
-		{
-			name: "Get not processed burn tickets not providing ethereum address, should not work",
-			body: func(t *testing.T) {
-				target := url.URL{Path: "/v1/client/get/not_processed_burn_tickets"}
-
-				query := target.Query()
-
-				query.Add("client_id", clientID)
-				query.Add("nonce", "0")
-
-				target.RawQuery = query.Encode()
-
-				req := httptest.NewRequest(http.MethodGet, target.String(), nil)
-
-				resp, err := serverChain.GetNotProcessedBurnTicketsHandler(context.Background(), req)
-				require.Error(t, err)
-				require.Nil(t, resp)
-			},
-		},
-		{
-			name: "Get not processed burn tickets not providing nonce, should work",
-			body: func(t *testing.T) {
-				err := eventDb.Get().Model(&event.BurnTicket{}).Create(&event.BurnTicket{
-					UserID:          clientID,
-					EthereumAddress: ethereumAddress,
-					Hash:            hash,
-					Nonce:           1,
-				}).Error
-				require.NoError(t, err)
-
-				target := url.URL{Path: "/v1/client/get/not_processed_burn_tickets"}
-
-				query := target.Query()
-
-				query.Add("ethereum_address", ethereumAddress)
-				query.Add("client_id", clientID)
-
-				target.RawQuery = query.Encode()
-
-				req := httptest.NewRequest(http.MethodGet, target.String(), nil)
-
-				respRaw, err := serverChain.GetNotProcessedBurnTicketsHandler(context.Background(), req)
-				require.NoError(t, err)
-
-				resp, ok := respRaw.([]*state.BurnTicket)
-				require.True(t, ok)
-				require.Len(t, resp, 1)
-
-				err = eventDb.Get().Model(&event.BurnTicket{}).Where("user_id = ?", clientID).Delete(&event.BurnTicket{}).Error
-				require.NoError(t, err)
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, test.body)
-	}
-}
-
 func TestChain_HandleSCRest_Status(t *testing.T) {
 	t.Skip("need to be reworked to work with new handler setup")
