+++ code/go/0chain.net/sharder/handler_test.go
@@ -1,14 +1,25 @@
 package sharder
 
 import (
+	"context"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"testing"
 
 	"0chain.net/chaincore/block"
+	"0chain.net/chaincore/config"
+	"0chain.net/chaincore/state"
 	"0chain.net/core/common"
 	"0chain.net/core/encryption"
+	"0chain.net/smartcontract/dbs/event"
+	"github.com/stretchr/testify/require"
+)
+
+const (
+	clientID        = "client_id"
+	ethereumAddress = "ethereum_address"
+	hash            = "hash"
 )
 
 func init() {
@@ -24,6 +35,173 @@ func makeTestURL(url url.URL, values map[string]string) string {
 	url.RawQuery = q.Encode()
 
 	return url.String()
+}
+
+func TestNotProcessedBurnTicketsHandler(t *testing.T) {
+	const baseUrl = "/v1/not_processed_burn_tickets"
+
+	b := block.NewBlock("", 10)
+	b.Hash = encryption.Hash("data")
+
+	sc := makeTestChain(t)
+	sc.AddBlock(b)
+	sc.SetLatestFinalizedBlock(b)
+
+	cl := initDBs(t)
+	defer cl()
+
+	eventDb, err := event.NewInMemoryEventDb(config.DbAccess{}, config.DbSettings{
+		Debug:                 true,
+		PartitionChangePeriod: 1,
+	})
+	require.NoError(t, err)
+
+	t.Cleanup(func() {
+		err = eventDb.Drop()
+		require.NoError(t, err)
+
+		eventDb.Close()
+	})
+
+	sc.EventDb = eventDb
+
+	tests := []struct {
+		name string
+		body func(t *testing.T)
+	}{
+		{
+			name: "Get not processed burn tickets of the client, which hasn't performed any burn operations, should work",
+			body: func(t *testing.T) {
+				target := url.URL{Path: "/v1/client/get/not_processed_burn_tickets"}
+
+				query := target.Query()
+
+				query.Add("ethereum_address", ethereumAddress)
+				query.Add("client_id", clientID)
+				query.Add("nonce", "0")
+
+				target.RawQuery = query.Encode()
+
+				req := httptest.NewRequest(http.MethodGet, target.String(), nil)
+
+				respRaw, err := NotProcessedBurnTicketsHandler(context.Background(), req)
+				require.NoError(t, err)
+
+				resp, ok := respRaw.([]*state.BurnTicket)
+				require.True(t, ok)
+				require.Len(t, resp, 0)
+			},
+		},
+		{
+			name: "Get not processed burn tickets of the client, which has performed burn operation, should work",
+			body: func(t *testing.T) {
+				err := eventDb.Get().Model(&event.BurnTicket{}).Create(&event.BurnTicket{
+					UserID:          clientID,
+					EthereumAddress: ethereumAddress,
+					Hash:            hash,
+					Nonce:           1,
+				}).Error
+				require.NoError(t, err)
+
+				target := url.URL{Path: "/v1/client/get/not_processed_burn_tickets"}
+
+				query := target.Query()
+
+				query.Add("ethereum_address", ethereumAddress)
+				query.Add("client_id", clientID)
+				query.Add("nonce", "0")
+
+				target.RawQuery = query.Encode()
+
+				req := httptest.NewRequest(http.MethodGet, target.String(), nil)
+
+				respRaw, err := NotProcessedBurnTicketsHandler(context.Background(), req)
+				require.NoError(t, err)
+
+				resp, ok := respRaw.([]*state.BurnTicket)
+				require.True(t, ok)
+				require.Len(t, resp, 1)
+
+				err = eventDb.Get().Model(&event.BurnTicket{}).Where("user_id = ?", clientID).Delete(&event.BurnTicket{}).Error
+				require.NoError(t, err)
+			},
+		},
+		{
+			name: "Get not processed burn tickets not providing client id, should not work",
+			body: func(t *testing.T) {
+				target := url.URL{Path: "/v1/client/get/not_processed_burn_tickets"}
+
+				query := target.Query()
+
+				query.Add("ethereum_address", ethereumAddress)
+				query.Add("nonce", "0")
+
+				target.RawQuery = query.Encode()
+
+				req := httptest.NewRequest(http.MethodGet, target.String(), nil)
+
+				resp, err := NotProcessedBurnTicketsHandler(context.Background(), req)
+				require.Error(t, err)
+				require.Nil(t, resp)
+			},
+		},
+		{
+			name: "Get not processed burn tickets not providing ethereum address, should not work",
+			body: func(t *testing.T) {
+				target := url.URL{Path: "/v1/client/get/not_processed_burn_tickets"}
+
+				query := target.Query()
+
+				query.Add("client_id", clientID)
+				query.Add("nonce", "0")
+
+				target.RawQuery = query.Encode()
+
+				req := httptest.NewRequest(http.MethodGet, target.String(), nil)
+
+				resp, err := NotProcessedBurnTicketsHandler(context.Background(), req)
+				require.Error(t, err)
+				require.Nil(t, resp)
+			},
+		},
+		{
+			name: "Get not processed burn tickets not providing nonce, should work",
+			body: func(t *testing.T) {
+				err := eventDb.Get().Model(&event.BurnTicket{}).Create(&event.BurnTicket{
+					UserID:          clientID,
+					EthereumAddress: ethereumAddress,
+					Hash:            hash,
+					Nonce:           1,
+				}).Error
+				require.NoError(t, err)
+
+				target := url.URL{Path: "/v1/client/get/not_processed_burn_tickets"}
+
+				query := target.Query()
+
+				query.Add("ethereum_address", ethereumAddress)
+				query.Add("client_id", clientID)
+
+				target.RawQuery = query.Encode()
+
+				req := httptest.NewRequest(http.MethodGet, target.String(), nil)
+
+				respRaw, err := NotProcessedBurnTicketsHandler(context.Background(), req)
+				require.NoError(t, err)
+
+				resp, ok := respRaw.([]*state.BurnTicket)
+				require.True(t, ok)
+				require.Len(t, resp, 1)
+
+				err = eventDb.Get().Model(&event.BurnTicket{}).Where("user_id = ?", clientID).Delete(&event.BurnTicket{}).Error
+				require.NoError(t, err)
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, test.body)
+	}
 }
 
 func TestBlockHandler(t *testing.T) {
